+++ CHANGELOG.md
@@ -26,6 +26,7 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
   `DUNE_THROW(ExceptionType) << a << b`.
 
 - Add deduction guides to `TupleVector` analogous to `std::tuple`.
+- Add concept definition `Std::indirectly_copyable` to constrain the `DenseMatrixAssigner`.
 
 ## Build system: Changelog